chore: move env setup to settings
